#394: Fix Augury hanging on load on Chrome Canary
Augury was not compatible with Chrome Canary, it would always give a
“Loading…” indicator. Some analysis had been done and posted to Trello
which speculated that this was a result of window.postMessage API
changing, but in fact it was totally unrelated to that. It turned out
to be a bug in Angular 2 which prevents recursive components from
working correctly with templateUrl. I will create a plunkr
demonstrating the issue and post the bug to the Angular project.

In the meantime, this change fixes Augury on Google Chrome.