Fixes #487 (#502)
Option elements are excluded from being added to the component tree.

However, we must also exclude them from being numbered when assigning
augury-id's.

Otherwise, if we have option elements with other elements after it, we
will get undefined elements in the parent component's child array in the
component tree, since it assigns to the child array using the augury-id
as an index to the array.

Example:
If we have 4 <option> elements with a <div> after it, then the four
<option> elements have augury-id 0 to 3 and the <div> would get id 4.

When adding to the child array of the parent, the four <option>
elements are not added. The <div> after is added, but is added to index
4 in the child array. This creates undefined elements in positions 0-3 in
the array.

The front-end is then given the component tree and a JS error is thrown
when the user highlights the null node.